Splitting utils (#11)
* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft

* chore: draft
