+++ .eslintignore
@@ -1,4 +1,5 @@
 dist/
 src/__tests__/
-jest.config.js
+src/utils/__tests__/
+jest.config.ts
 webpack.config.js
