+++ src/defaultOptions.ts
@@ -1,0 +1,6 @@
+import type { AxiosAuthRefreshOptions } from './model';
+
+export const defaultOptions: AxiosAuthRefreshOptions = {
+    statusCodes: [401],
+    pauseInstanceWhileRefreshing: false,
+};
