+++ src/utils/__tests__/unsetCache.spec.ts
@@ -1,6 +1,6 @@
-import type { AxiosAuthRefreshCache } from '../model';
-import { mockedAxios } from './testUtil';
-import { unsetCache } from '../utils';
+import type { AxiosAuthRefreshCache } from '../../model';
+import { unsetCache } from '../unsetCache';
+import { axiosMock } from '../../test/axiosMock';
 
 describe('State is cleared', () => {
     const cache: AxiosAuthRefreshCache = {
@@ -10,12 +10,12 @@ describe('State is cleared', () => {
     };
 
     it('after refreshing call succeeds/fails', () => {
-        const instance = mockedAxios();
+        const instance = axiosMock();
         cache.requestQueueInterceptorId = instance.interceptors.request.use(() => undefined);
-        cache.skipInstances.push(instance);
+        cache.skipInstances.push(instance as any);
         expect(instance.interceptors.has('request', cache.requestQueueInterceptorId)).toBeTruthy();
         expect(cache.skipInstances.length).toBe(1);
-        unsetCache(instance, cache);
+        unsetCache(instance as any, cache);
         expect(cache.skipInstances.length).toBe(0);
         expect(cache.requestQueueInterceptorId).toBeFalsy();
         expect(instance.interceptors.has('request', cache.requestQueueInterceptorId)).toBeFalsy();
