+++ tsconfig.json
@@ -11,5 +11,5 @@
         "noImplicitReturns": true,
         "noFallthroughCasesInSwitch": true
     },
-    "exclude": ["node_modules", "dist", "jest.config.js", "webpack.config.js"]
+    "exclude": ["node_modules", "dist", "jest.config.ts", "webpack.config.js"]
 }
